refactored projects code
