+++ src/app/projects/types.tsx
@@ -1,0 +1,6 @@
+export default interface ProjectPage {
+    title: string;
+    id: string;
+    content: string;
+    url: string;
+}
